feat: Add fibonacci series generator
This commit adds code to generate the fibonacci series of numbers based on user input. The fibonacci function takes an integer input and prints the fibonacci series up to that number. The user is prompted to enter a number between 1 and 100. After generating the series, a thank you message is displayed.